# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API da Teste Receitas RG Sistemas",
        "description": "Documentação da API para o projeto da Teste Receitas RG Sistemas",
        "contact": {
            "email": "rafa.jefer@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/recipes/{id}": {
            "put": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Atualiza uma receita",
                "operationId": "b88b977d98ca0ff193538bdc4325bea7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "preparation_time_minutes": {
                                        "type": "integer"
                                    },
                                    "servings": {
                                        "type": "integer"
                                    },
                                    "ingredients": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "steps": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "title": "Bolo de Cenoura modificado",
                                "description": "Um bolo clássico, fofinho e delicioso.",
                                "ingredients": [
                                    "3 cenouras médias raladas",
                                    "4 ovos",
                                    "1 xícara de óleo",
                                    "2 xícaras de açúcar",
                                    "2 e 1/2 xícaras de farinha de trigo",
                                    "1 colher de sopa de fermento em pó"
                                ],
                                "steps": [
                                    "Bata no liquidificador as cenouras, ovos e óleo.",
                                    "Adicione o açúcar e bata novamente.",
                                    "Misture a farinha e o fermento em uma tigela e acrescente a mistura do liquidificador.",
                                    "Despeje em uma forma untada e asse em forno médio por 40 minutos."
                                ],
                                "preparation_time_minutes": 60,
                                "servings": 10,
                                "category_id": null
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receita atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Bolo de Cenoura modificado"
                                        },
                                        "preparationTimeMinutes": {
                                            "type": "integer",
                                            "example": null
                                        },
                                        "servings": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "ingredients": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "steps": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "user_id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "id": "1",
                                    "title": "Bolo de Cenoura modificado",
                                    "preparationTimeMinutes": null,
                                    "servings": 10,
                                    "ingredients": [
                                        "3 cenouras médias raladas",
                                        "4 ovos",
                                        "1 xícara de óleo",
                                        "2 xícaras de açúcar",
                                        "2 e 1/2 xícaras de farinha de trigo",
                                        "1 colher de sopa de fermento em pó"
                                    ],
                                    "steps": [
                                        "Bata no liquidificador as cenouras, ovos e óleo.",
                                        "Adicione o açúcar e bata novamente.",
                                        "Misture a farinha e o fermento em uma tigela e acrescente a mistura do liquidificador.",
                                        "Despeje em uma forma untada e asse em forno médio por 40 minutos."
                                    ],
                                    "user_id": "1",
                                    "category_id": null
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Remove uma receita",
                "operationId": "3a5b52369b6fd6ea1411207b8d2ea876",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receita removida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Receita removida com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recipes/{id}/print": {
            "get": {
                "tags": [
                    "Receitas"
                ],
                "summary": "Imprime uma receita em PDF",
                "operationId": "fa47e84438dadf6ac3d66ab5e7a71f1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF gerado com sucesso",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}